Refactor sbatch_args handling: move functions to sbatch_args.py and update imports
